+++ dialogExecution/editVMNew.py
@@ -18,6 +18,9 @@ import translations.en
 import translations.fr
 import translations.es
 import translations.ro
+import translations.be
+import translations.cz
+import translations.ru
 import locale
 
 class EditVMNewDialog(QDialog, Ui_Dialog):
@@ -91,6 +94,15 @@ class EditVMNewDialog(QDialog, Ui_Dialog):
                 elif result[0][1] == "ro":
                     langmode = "ro"
 
+                elif result[0][1] == "ru":
+                    langmode = "ru"
+
+                elif result[0][1] == "be":
+                    langmode = "be"
+
+                elif result[0][1] == "cz":
+                    langmode = "cz"
+
                 elif result[0][1] == "system":
                     langmode = "system"
 
@@ -130,6 +142,15 @@ class EditVMNewDialog(QDialog, Ui_Dialog):
             elif languageToUse.startswith("ro"):
                 translations.ro.translateEditVMRO(self)
 
+            elif languageToUse.startswith("ru"):
+                translations.ru.translateEditVMRU(self)
+
+            elif languageToUse.startswith("be"):
+                translations.be.translateEditVMBE(self)
+
+            elif languageToUse.startswith("cz"):
+                translations.cz.translateEditVMCZ(self)
+
             else:
                 translations.en.translateEditVMEN(self)
         
@@ -160,6 +181,15 @@ class EditVMNewDialog(QDialog, Ui_Dialog):
 
                     elif languageToUse.startswith("ro"):
                         translations.ro.translateEditVMRO(self)
+
+                    elif languageToUse.startswith("ru"):
+                        translations.ru.translateEditVMRU(self)
+
+                    elif languageToUse.startswith("be"):
+                        translations.be.translateEditVMBE(self)
+
+                    elif languageToUse.startswith("cz"):
+                        translations.cz.translateEditVMCZ(self)
 
                     else:
                         translations.en.translateEditVMEN(self)