+++ dialogExecution/newVirtualMachine.py
@@ -18,6 +18,9 @@ import translations.en
 import translations.fr
 import translations.es
 import translations.ro
+import translations.be
+import translations.cz
+import translations.ru
 import locale
 
 class NewVirtualMachineDialog(QDialog, Ui_Dialog):
@@ -174,6 +177,15 @@ class NewVirtualMachineDialog(QDialog, Ui_Dialog):
                 elif result[0][1] == "ro":
                     langmode = "ro"
 
+                elif result[0][1] == "ru":
+                    langmode = "ru"
+
+                elif result[0][1] == "be":
+                    langmode = "be"
+
+                elif result[0][1] == "cz":
+                    langmode = "cz"
+
                 elif result[0][1] == "system":
                     langmode = "system"
 
@@ -211,6 +223,15 @@ class NewVirtualMachineDialog(QDialog, Ui_Dialog):
             elif languageToUse.startswith("ro"):
                 translations.ro.translateNewVmRO(self)
 
+            elif languageToUse.startswith("ru"):
+                translations.ru.translateNewVmRU(self)
+
+            elif languageToUse.startswith("be"):
+                translations.be.translateNewVmBE(self)
+
+            elif languageToUse.startswith("cz"):
+                translations.cz.translateNewVmCZ(self)
+
             else:
                 translations.en.translateNewVmEN(self)
         
@@ -241,6 +262,15 @@ class NewVirtualMachineDialog(QDialog, Ui_Dialog):
 
                     elif languageToUse.startswith("ro"):
                         translations.ro.translateNewVmRO(self)
+
+                    elif languageToUse.startswith("ru"):
+                        translations.ru.translateNewVmRU(self)
+
+                    elif languageToUse.startswith("be"):
+                        translations.be.translateNewVmBE(self)
+
+                    elif languageToUse.startswith("cz"):
+                        translations.cz.translateNewVmCZ(self)
 
                     else:
                         translations.en.translateNewVmEN(self)