+++ dialogExecution/qemuSysError.py
@@ -15,6 +15,9 @@ import translations.en
 import translations.fr
 import translations.es
 import translations.ro
+import translations.be
+import translations.cz
+import translations.ru
 import locale
 import sqlite3
 
@@ -102,6 +105,15 @@ class QemuSysMissing(QDialog, Ui_Dialog):
                 elif result[0][1] == "ro":
                     langmode = "ro"
 
+                elif result[0][1] == "ru":
+                    langmode = "ru"
+
+                elif result[0][1] == "be":
+                    langmode = "be"
+
+                elif result[0][1] == "cz":
+                    langmode = "cz"
+
                 elif result[0][1] == "system":
                     langmode = "system"
 
@@ -139,6 +151,15 @@ class QemuSysMissing(QDialog, Ui_Dialog):
             elif languageToUse.startswith("ro"):
                 translations.ro.translateQemuSysMissingRO(self, self.vmSpecs[1])
 
+            elif languageToUse.startswith("ru"):
+                translations.ru.translateQemuSysMissingRU(self, self.vmSpecs[1])
+
+            elif languageToUse.startswith("be"):
+                translations.be.translateQemuSysMissingBE(self, self.vmSpecs[1])
+
+            elif languageToUse.startswith("cz"):
+                translations.cz.translateQemuSysMissingCZ(self, self.vmSpecs[1])
+
             else:
                 translations.en.translateQemuSysMissingEN(self, self.vmSpecs[1])
         
@@ -169,6 +190,15 @@ class QemuSysMissing(QDialog, Ui_Dialog):
 
                     elif languageToUse.startswith("ro"):
                         translations.ro.translateQemuSysMissingRO(self, self.vmSpecs[1])
+
+                    elif languageToUse.startswith("ru"):
+                        translations.ru.translateQemuSysMissingRU(self, self.vmSpecs[1])
+
+                    elif languageToUse.startswith("be"):
+                        translations.be.translateQemuSysMissingBE(self, self.vmSpecs[1])
+
+                    elif languageToUse.startswith("cz"):
+                        translations.cz.translateQemuSysMissingCZ(self, self.vmSpecs[1])
 
                     else:
                         translations.en.translateQemuSysMissingEN(self, self.vmSpecs[1])