+++ dialogExecution/updateAvailable.py
@@ -14,6 +14,9 @@ import sqlite3
 import translations.de
 import translations.uk
 import translations.en
+import translations.be
+import translations.cz
+import translations.ru
 import locale
 
 class UpdateAvailable(QDialog, Ui_Dialog):