+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# EmuGUI v1.1.0.5401_dev "Sara Angeline" (based on v1.1.0.5400_dev)
+
+## Changes compared to v1.0.1.5311
+
+- New feature level, new codename. I decided to call it "Sara Angeline" based on what I want to add so far.
+- Language files for French, Spanish, Czech, Russian, Belarusian and Romanian have been prepared.
+- Thanks for translating EmuGUI into Russian, Danik2343.
+
 # EmuGUI v1.1.0.5400_dev "Sara Angeline" (based on v1.0.1.5311)
 
 ## Changes compared to v1.0.1.5311